aries-common: Support systemless SuperSU
Manually patch the bootimage and sepolicy (requires patch to external/sepolicy)

Only used when the flag "USE_SUPERSU := true" is added to the BoardConfig.mk due to the selinux changes (but really, for now, as we're in permissive mode there's no reason not to include the changes).

If a user hasn't flashed the SuperSU, then the only changes are that /su is created and the path variable is modified.

Tested only with SuperSU 2.71.  Other versions may or may not work

Conflicts:
	fstab.aries
	init.aries.rc
	patches/apply_patches.sh